RawData_Check/djangoverifier/backend/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
blinker==1.6.2
cffi==1.15.1
chardet==5.1.0
click==8.1.3
cryptography==41.0.1
Django==4.2.2
django-cache-url==3.4.4
django-cors-headers==4.1.0
djangorestframework==3.14.0
dnspython==2.3.0
et-xmlfile==1.1.0
Flask==2.3.2
future==0.18.3
impacket==0.10.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
ldap3==2.9.1
ldapdomaindump==0.9.4
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
numpy==1.25.0
openpyxl==3.1.2
pandas==2.0.2
paramiko==3.2.0
protobuf==3.20.3
pyasn1==0.5.0
pycparser==2.21
pycryptodomex==3.18.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pysmb==1.2.9.1
python-dateutil==2.8.2
python-decouple==3.8
python-memcached==1.59
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
Werkzeug==2.3.6
zipp==3.15.0
